+++ src/components/feed-sources/feed-sources-columns.jsx
@@ -1,9 +1,9 @@
 import { React, useContext } from "react";
 import { useTranslation } from "react-i18next";
 import ColumnHoc from "../util/column-hoc";
-import ListButton from "../util/list/list-button.jsx";
-import SelectColumnHoc from "../util/select-column-hoc.jsx";
-import UserContext from "../../context/user-context.jsx";
+import ListButton from "../util/list/list-button";
+import SelectColumnHoc from "../util/select-column-hoc";
+import UserContext from "../../context/user-context";
 
 /**
  * Retrieves the columns for feed sources data based on API call response.
@@ -12,7 +12,6 @@ import UserContext from "../../context/user-context.jsx";
  * @param {Function} props.apiCall - The API call function to retrieve feed sources data.
  * @param {string} props.infoModalRedirect - The redirect URL for information modal.
  * @param {string} props.infoModalTitle - The title for information modal.
- * @param {string} props.dataKey - The key for data retrieval.
  * @returns {object} Columns - An array of objects representing the columns for
  *   feed sources data.
  */