+++ src/remote-component.config.js
@@ -1,5 +1,6 @@
 import react from "react";
 
+// eslint-disable-next-line import/prefer-default-export
 export const resolve = {
   react,
 };