+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#273](https://github.com/os2display/display-admin-client/pull/273)
+  - Fixed calendar api feed source config endpoint.
 - [#272](https://github.com/os2display/display-admin-client/pull/272)
   - Fixed feed source selector.
 - [#264](https://github.com/os2display/display-admin-client/pull/264)