+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-- [#TODO](https://github.com/os2display/display-admin-client/pull/TODO)
+- [#274](https://github.com/os2display/display-admin-client/pull/274)
   - Added screen status to screen list.
   - Refactored screen status on screen edit.
   - Change campaign icon in screen list to boolean text.
+- [#273](https://github.com/os2display/display-admin-client/pull/273)
+  - Fixed calendar api feed source config endpoint.
+- [#272](https://github.com/os2display/display-admin-client/pull/272)
+  - Fixed feed source selector.
 - [#264](https://github.com/os2display/display-admin-client/pull/264)
   - Added checkbox options component for use in calendar modifiers.
   - Fixed multiselect when more than one feed source of the given type is installed.